webapp/requirements.txt
django
django-heroku
requests
gunicorn
gitpython
pandas
sklearn
